TOPU-6 feat: Implement UnAuthentication exception
